requirements.txt
# 图像处理相关
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# AI 水印去除模型
iopaint>=1.3.0
